chore: deploy marketplace script

+++ modules/tool/api/upload/confirmUpload.ts
@@ -2,24 +2,56 @@ import { s } from '@/router/init';
 import { contract } from '@/contract';
 import { MongoS3TTL } from '@/s3/ttl/schema';
 import { UploadToolsS3Path } from '@tool/constants';
-import { MongoPlugin } from '@/mongo/models/plugins';
+import { MongoPlugin, pluginTypeEnum, PluginZodSchema } from '@/mongo/models/plugins';
 import { refreshVersionKey } from '@/cache';
 import { SystemCacheKeyEnum } from '@/cache/type';
+import { mongoSessionRun } from '@/mongo/utils';
 
 export default s.route(contract.tool.upload.confirmUpload, async ({ body }) => {
   const { toolIds } = body;
 
-  await MongoPlugin.create(
-    toolIds.map((toolId) => ({
-      toolId,
-      type: 'tool'
-    }))
-  );
+  const tools = await MongoPlugin.find({
+    toolId: {
+      $in: toolIds
+    },
+    type: pluginTypeEnum.Enum.tool
+  });
 
-  await MongoS3TTL.deleteMany({
-    minioKey: {
-      $regex: toolIds.map((toolId) => `^${UploadToolsS3Path}/${toolId}`).join('|')
-    }
+  if (tools.length !== toolIds.length) {
+    return {
+      status: 400,
+      body: {
+        message: 'Some toolIds are invalid'
+      }
+    };
+  }
+
+  await mongoSessionRun(async (session) => {
+    await MongoPlugin.updateMany(
+      {
+        toolId: {
+          $in: toolIds
+        }
+      },
+      {
+        $set: {
+          status: PluginZodSchema.shape.status.Enum.active
+        }
+      },
+      {
+        session
+      }
+    );
+    await MongoS3TTL.deleteMany(
+      {
+        minioKey: {
+          $regex: toolIds.map((toolId) => `^${UploadToolsS3Path}/${toolId}`).join('|')
+        }
+      },
+      {
+        session
+      }
+    );
   });
 
   await refreshVersionKey(SystemCacheKeyEnum.systemTool);
